Refactored common model-checking helpers and types out of `bmc.rs`
