+++ patronus/src/mc.rs
@@ -5,10 +5,11 @@
 mod bmc;
 mod pdr;
 mod types;
+mod encoding;
+mod utils;
 
-pub use bmc::{
-    ModelCheckResult, TransitionSystemEncoding, UnrollSmtEncoding, bmc, check_assuming,
-    check_assuming_end, get_smt_value,
-};
+pub use bmc::bmc;
 pub use pdr::pdr;
-pub use types::{InitValue, Witness};
+pub use types::{InitValue, Witness, ModelCheckResult};
+pub use encoding::{TransitionSystemEncoding, UnrollSmtEncoding};
+pub use utils::{check_assuming, check_assuming_end, get_smt_value};
