+++ patronus/src/mc/pdr.rs
@@ -4,8 +4,9 @@
 
 use crate::expr::*;
 use crate::mc::bmc::start_bmc_or_pdr;
+use crate::mc::encoding::TransitionSystemEncoding;
 use crate::mc::{
-    ModelCheckResult, TransitionSystemEncoding, bmc, check_assuming, check_assuming_end,
+    ModelCheckResult, bmc, check_assuming, check_assuming_end,
     get_smt_value,
 };
 use crate::smt::*;
